Switch Windows tests to Clang

+++ native/crypto/Random.cc
@@ -5,7 +5,7 @@
 #include <random>
 
 void Random::Get(void *data, size_t size) {
-    static std::independent_bits_engine<std::random_device, CHAR_BIT, u8> engine;
+    static std::independent_bits_engine<std::random_device, CHAR_BIT, u16> engine;
     std::generate(reinterpret_cast<u8 *>(data), reinterpret_cast<u8 *>(data) + size,
             std::ref(engine));
 }
